chore: move parse csv out of s3 client, into its own service
