+++ backend/src/core/services/index.ts
@@ -7,3 +7,4 @@ export * from './credential.service'
 export * from './stats.service'
 export * from './api-key.service'
 export * from './auth.service'
+export * from './parse-csv.service'
